server/requirements.txt
python-dotenv
PyMuPDF
openai
pdf2docx
flask-restful
Flask-Cors
fpdf2
cmake
